requirements.txt
Django>=3.0,<4.0
Django
djangorestframework
psycopg2
django-environ